+++ src/db/models/users.js
@@ -164,7 +164,7 @@ Users.prototype.comparePassword = async function (password) {
   return await bcrypt.compare(password, this.passwordHash);
 };
 
-Users.encryptPassword = async password => await bcrypt.hash(password, saltRounds);
+Users.encryptPassword = password => bcrypt.hash(password, saltRounds);
 
 Users.isOfLegalAge = bornDateString => {
   const bornDate = parseISO(bornDateString);
@@ -218,11 +218,6 @@ Users.beforeCreate(async (user, options) => {
 // const locations = await user.getLocations();
 // if (locations) {
 // await locations.destroy();
-// }
-
-// const phones = await user.getPhones();
-// if (phones) {
-// await phones.destroy();
 // }
 // });